+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * **Inspection**: *property without annotation* renamed to *column without annotation*;
 
+### Fixed
+
+* **NPE**: on skip parenteshes, consider that it could have any argument yet (eg. `()->id`);
+
 ## 0.1.0 [[Monogon Release](https://github.com/rentalhost/laravel-insight/releases/tag/0.1.0)] - 2017-05-31
 
 ### Added