+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.0 - [Unreleased]
+
+### Added
+
+* **Inspection**: *column without annotation* now try to guess the column type (read the [doc](https://github.com/rentalhost/laravel-insight/wiki/Inspections#column-without-annotation) to more information);
+
 ## 0.1.1 - 2017-06-01
 
 ### Changed